qdsplit option -t can now be used to speed up splitting large files
Querydata has the time dimension as the innermost variable. Hence
looping over the data with the time dimensions in the outermost
loop is inefficient, and for very large data can lead to cache
trashing.

The new -t option can be used to set how many time steps are
extracted simultaneously. For very large files the value might
be for example one half or one third of the total number of
time steps. The value must be tuned based on the amount of
memory required to hold the time steps simultaneously in memory
as well as much how much cache trashing will occur if the
input data is iterated over too many times.