Changed default centre for qdtogrib to be unset
Setting the default producer to be EFKL changes the default parameter
conversion tables. Unfortunately WindDirection is not defined in the EFKL
tables and hence converting the data produced by the SmartMet Editor
failed.

The default producer is now empty meaning the default tables will be used.
The subcentre option now has no effect unless the centre option is used.